requirements.txt
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
Pillow==10.0.0
requests==2.31.0
django-twilio==0.14.3.2
django-cors-headers==4.2.0
gunicorn==20.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6